feat(): support to find references of a template
