+++ bricks/next-builder/src/index.ts
@@ -14,3 +14,4 @@ import "./data-providers/GetAllProviders";
 import "./data-providers/GetBrickLibrary";
 import "./data-providers/StoryboardAssembly";
 import "./data-providers/ScanI18nInStoryboard";
+import "./data-providers/FindReferencesOfTemplate";
